requirements.txt
mcdreforged>=2.2.0
python-socketio
